[Feat] #9 친구추가 구현완료(Controller)
