+++ src/main/java/com/photodiary/backend/friend/Exception/CannotAddYourselfAsFriendException.java
@@ -1,0 +1,7 @@
+package com.photodiary.backend.friend.Exception;
+
+public class CannotAddYourselfAsFriendException extends RuntimeException {
+    public CannotAddYourselfAsFriendException(String message) {
+        super(message);
+    }
+}
