+++ src/main/java/com/photodiary/backend/friend/dto/AddFriendRequestDto.java
@@ -1,0 +1,8 @@
+package com.photodiary.backend.friend.dto;
+
+import lombok.Getter;
+
+@Getter
+public class AddFriendRequestDto {
+    String email;
+}
